Loading state UI and error handling.
Added a loading spinner to the middle of the screen when the paging library is performing a Refresh load. I also added a header and footer to the Browse recycler view that display a progress bar/error message/retry button when a append/prepend load or load error occurs.

Added network error handling to the RxJava network call in the remote mediator.